Removed DefaultButtons and fonts function, TODO createDropDown function
Put the defaults inline with var fomts = options or default array.
instead of calling a function that ruturns the same array.

We need to make a create drop down function that will take the buttons
name and HTML to insert in the button as paramemters and then append
that html to the button.