Added createDropDown function, fixed some errors.
Added a createDropDown function. This now accepts HTML and inserts it
into your button.
!Fixed dropdown toggle when clicking inside the dropdown to prevent it
closing if its not a selection.

Todo: Add an option to detect if submit button was selected in a
dropdown.

Pass buttons to functions based on their css class name.  custom
buttons.

detect if empty and and a <div> to stop it from deleting.
Can do this on keypress del or backspace.
detect if parent is body, to prevent it deleting the first div also on
backspace.

Parse BBC.

Add plugin scripts to header.